Refactoring bankroll to better module structure
