+++ app/build.gradle
@@ -49,7 +49,9 @@ dependencies {
     implementation project(path: ':features:tournaments:presentation')
     implementation project(path: ':features:bankroll:data')
     implementation project(path: ':features:bankroll:domain')
-    implementation project(path: ':features:bankroll:presentation')
+    implementation project(path: ':features:bankroll:presentation:creation')
+    implementation project(path: ':features:bankroll:presentation:navigation')
+    implementation project(path: ':features:bankroll:presentation:list')
     implementation project(path: ':features:grind:data')
     implementation project(path: ':features:grind:domain')
     implementation project(path: ':features:grind:presentation')
