+++ features/bankroll/presentation/balance-component/src/main/AndroidManifest.xml
