--- features/bankroll/presentation/build.gradle
@@ -1,10 +1,0 @@
-apply from: "$rootDir/config/module/feature/presentation.gradle"
-
-dependencies {
-    // Internal modules
-    implementation project(path: ':features:bankroll:domain')
-    implementation project(path: ':features:bankroll:components:balance')
-
-    // Ui
-    implementation "androidx.cardview:cardview:1.0.0"
-}
