+++ features/bankroll/presentation/creation/build.gradle
@@ -1,0 +1,10 @@
+apply from: "$rootDir/config/module/feature/presentation.gradle"
+
+dependencies {
+    // Internal modules
+    implementation project(path: ':features:bankroll:domain')
+    implementation project(path: ':features:bankroll:presentation:navigation')
+
+    // Ui
+    implementation "androidx.cardview:cardview:1.0.0"
+}
