+++ features/bankroll/presentation/list/src/main/res/layout/fragment_bankroll_item.xml
