+++ features/bankroll/presentation/navigation/build.gradle
@@ -1,0 +1,6 @@
+apply from: "$rootDir/config/module/feature/presentation.gradle"
+
+dependencies {
+    // Internal modules
+    implementation project(path: ':features:bankroll:domain')
+}
