+++ features/bankroll/presentation/navigation/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest package="com.mctech.pokergrinder.bankroll.presentation.navigation" />
