+++ features/bankroll/presentation/navigation/src/main/java/com/mctech/pokergrinder/bankroll/presentation/navigation/BankrollNavigation.kt
@@ -1,0 +1,6 @@
+package com.mctech.pokergrinder.bankroll.presentation.navigation
+
+public interface BankrollNavigation {
+  public fun goToBankrollDeposit()
+  public fun navigateBack()
+}
