--- features/bankroll/presentation/src/main/java/com/mctech/pokergrinder/bankroll/presentation/list/BankrollNavigation.kt
@@ -1,6 +1,0 @@
-package com.mctech.pokergrinder.bankroll.presentation.list
-
-public interface BankrollNavigation {
-  public fun goToBankrollDeposit()
-  public fun navigateBack()
-}
