+++ features/summary/presentation/build.gradle
@@ -2,7 +2,7 @@ apply from: "$rootDir/config/module/feature/presentation.gradle"
 
 dependencies {
     // Internal modules
-    implementation project(path: ':features:bankroll:components:balance')
+    implementation project(path: ':features:bankroll:presentation:balance-component')
     implementation project(path: ':features:summary:domain')
     implementation project(path: ':features:summary:components:session')
     implementation project(path: ':features:summary:components:investment')
